pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recommendation-system"
version = "0.10"
description = "Built on Coursera scraper data. Models: KNN, Baseline, NN (GPU only). CLI tool to train and save models."
readme = "ReadMe.md"
license = { text = "MIT" }
authors = [{ name = "cmd-HMN", email = "hafizmuhammadnoor01@gmail.com" }]
requires-python = ">=3.12"
keywords = ["recommendation", "machine-learning", "CLI"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux"
]
dependencies = [
    "flake8>=7.0",
    "numpy>=1.27",
    "pandas>=2.1",
    "scikit-learn>=2.2",
    "torch>=2.2",
    "scikit-surprise>=1.1.3"
]

[project.scripts]
recommendation-cli = "cli:main"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E501", "E722", "F401"]
